requirements-dev.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
autopep8==1.4.4
Click==7.0
Django==3.0.2
django-filter==2.2.0
djangorestframework==3.11.0
pathspec==0.7.0
pycodestyle==2.5.0
python-decouple==3.3
pytz==2019.3
regex==2020.1.8
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
